+++ mtools/mlaunch/mlaunch.py
@@ -7,6 +7,8 @@ import os, time
 import socket
 import json
 
+from mtools.util.cmdlinetool import BaseCmdLineTool
+
 try:
     from pymongo import Connection
     from pymongo.errors import ConnectionFailure, AutoReconnect, OperationFailure
@@ -15,6 +17,7 @@ except ImportError:
 
 
 def pingMongoDS(host, interval=1, timeout=30):
+    """ Ping a mongos or mongod every `interval` seconds until it responds, or `timeout` seconds have passed. """
     con = None
     startTime = time.time()
     while not con:
@@ -27,54 +30,55 @@ def pingMongoDS(host, interval=1, timeout=30):
             time.sleep(interval)
 
 
-class MongoLauncher(object):
+class MLaunchTool(BaseCmdLineTool):
+
+    # additional arguments that can be passed on to mongos, all others will only go to mongod
+    mongos_arguments = ['quiet', 'logappend', 'pidfilepath', 'keyFile', 'nounixsocket', 'syslog', \
+                        'nohttpinterface', 'test', 'upgrade', 'ipv6', 'jsonp', 'noscripting']
 
     def __init__(self):
+        BaseCmdLineTool.__init__(self)
+        self.argparser.description = 'script to launch MongoDB stand-alone servers, replica sets and shards'
 
-        self.parseArgs()
+        # positional argument
+        self.argparser.add_argument('dir', action='store', nargs='?', const='.', default='.', help='base directory to create db and log paths (default=.)')
+
+        # either single or replica set
+        me_group = self.argparser.add_mutually_exclusive_group(required=True)
+        me_group.add_argument('--single', action='store_true', help='creates a single stand-alone mongod instance')
+        me_group.add_argument('--replicaset', action='store_true', help='creates replica set with several mongod instances')
+        me_group.add_argument('--restore', action='store_true', help='restores a previously launched existing configuration from the data directory.')
+
+        self.argparser.add_argument('--nodes', action='store', metavar='NUM', type=int, default=3, help='adds NUM data nodes to replica set (requires --replicaset, default=3)')
+        self.argparser.add_argument('--arbiter', action='store_true', default=False, help='adds arbiter to replica set (requires --replicaset)')
+        self.argparser.add_argument('--name', action='store', metavar='NAME', default='replset', help='name for replica set (default=replset)')
+        
+        # sharded or not
+        self.argparser.add_argument('--sharded', action='store', nargs='*', metavar='N', help='creates a sharded setup consisting of several singles or replica sets. Provide either list of shard names or number of shards (default=1)')
+        self.argparser.add_argument('--config', action='store', default=1, type=int, metavar='NUM', choices=[1, 3], help='adds NUM config servers to sharded setup (requires --sharded, NUM must be 1 or 3, default=1)')
+        self.argparser.add_argument('--mongos', action='store', default=1, type=int, metavar='NUM', help='starts NUM mongos processes (requires --sharded, default=1)')
+
+        # verbose, port, auth, loglevel
+        self.argparser.add_argument('--verbose', action='store_true', default=False, help='outputs information about the launch')
+        self.argparser.add_argument('--port', action='store', type=int, default=27017, help='port for mongod, start of port range in case of replica set or shards (default=27017)')
+        self.argparser.add_argument('--authentication', action='store_true', default=False, help='enable authentication and create a key file and admin user (admin/mypassword)')
+        self.argparser.add_argument('--rest', action='store_true', default=False, help='enable REST interface on mongod processes')
+        self.argparser.add_argument('--local', action='store_true', default=False, help='run mongod/s process from local directory, i.e. "./mongod"')
+
         self.hostname = socket.gethostname()
 
+
+    def run(self):
+        # don't call BaseCmdLineTool.run(), we pass args ourselves
+        self.args, self.unknown_args = self.argparser.parse_known_args()
+        self.args = vars(self.args)
+
+        # load or store parameters
         if self.args['restore']:
             self.load_parameters()
         else:
             self.store_parameters()
 
-        self.launch()
-
-    def parseArgs(self):
-        # create parser object
-        parser = argparse.ArgumentParser(description='script to launch MongoDB stand-alone servers, replica sets, and shards')
-        
-        # positional argument
-        parser.add_argument('dir', action='store', nargs='?', const='.', default='.', help='base directory to create db and log paths')
-
-        # either single or replica set
-        me_group = parser.add_mutually_exclusive_group(required=True)
-        me_group.add_argument('--single', action='store_true', help='creates a single stand-alone mongod instance')
-        me_group.add_argument('--replicaset', action='store_true', help='creates replica set with several mongod instances')
-        me_group.add_argument('--restore', action='store_true', help='restores a previously launched existing configuration from the data directory.')
-
-        parser.add_argument('--nodes', action='store', metavar='NUM', type=int, default=3, help='adds NUM data nodes to replica set (requires --replicaset, default: 3)')
-        parser.add_argument('--arbiter', action='store_true', default=False, help='adds arbiter to replica set (requires --replicaset)')
-        parser.add_argument('--name', action='store', metavar='NAME', default='replset', help='name for replica set (default: replset)')
-        
-        # sharded or not
-        parser.add_argument('--sharded', action='store', nargs='*', metavar='N', help='creates a sharded setup consisting of several singles or replica sets. Provide either list of shard names or number of shards (default: 1)')
-        parser.add_argument('--config', action='store', default=1, type=int, metavar='NUM', choices=[1, 3], help='adds NUM config servers to sharded setup (requires --sharded, NUM must be 1 or 3, default: 1)')
-
-        # verbose, port, auth, loglevel
-        parser.add_argument('--verbose', action='store_true', default=False, help='outputs information about the launch')
-        parser.add_argument('--port', action='store', type=int, default=27017, help='port for mongod, start of port range in case of replica set or shards (default: 27017)')
-        parser.add_argument('--authentication', action='store_true', default=False, help='enable authentication and create a key file and admin user (admin/mypassword)')
-        parser.add_argument('--loglevel', action='store', default=False, type=int, help='increase loglevel to LOGLEVEL (default: 0)')
-        parser.add_argument('--rest', action='store_true', default=False, help='enable REST interface on mongod processes')
-        parser.add_argument('--local', action='store_true', default=False, help='run mongod/s process from local directory, i.e. "./mongod"')
-        self.args = vars(parser.parse_args())
-        if self.args['verbose']:
-            print "parameters:", self.args
-
-
-    def launch(self):
         datapath = os.path.join(self.args['dir'], 'data')
         
         # check if authentication is enabled        
@@ -87,9 +91,9 @@ class MongoLauncher(object):
         if self.args['sharded']:
             self._launchSharded()
         elif self.args['single']:
-            self._launchSingle(self.args['dir'], self.args['port'], verbose=self.args['verbose'])
+            self._launchSingle(self.args['dir'], self.args['port'])
         elif self.args['replicaset']:
-            self._launchReplSet(self.args['dir'], self.args['port'], self.args['name'], self.args['nodes'], self.args['arbiter'], self.args['verbose'])
+            self._launchReplSet(self.args['dir'], self.args['port'], self.args['name'])
 
 
     def convert_u2b(self, obj):
@@ -121,15 +125,8 @@ class MongoLauncher(object):
         except Exception:
             pass
 
-    def check_port_availability(self, port, binary):
-        s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        try:
-            s.bind(('', port))
-            s.close()
-        except:
-            exit("Can't start " + binary + ", port " + str(port) + " is already being used")
-
-    def _createPaths(self, basedir, name=None, verbose=False):
+    
+    def _createPaths(self, basedir, name=None):
         if name:
             datapath = os.path.join(basedir, 'data', name)
         else:
@@ -138,7 +135,7 @@ class MongoLauncher(object):
         dbpath = os.path.join(datapath, 'db')
         if not os.path.exists(dbpath):
             os.makedirs(dbpath)
-        if verbose:
+        if self.args['verbose']:
             print 'creating directory: %s'%dbpath
         
         return datapath
@@ -162,10 +159,10 @@ class MongoLauncher(object):
         nextport = self.args['port']
         for p, shard in enumerate(shard_names):
             if self.args['single']:
-                shard_names[p] = self._launchSingle(self.args['dir'], nextport, name=shard, verbose=self.args['verbose'])
+                shard_names[p] = self._launchSingle(self.args['dir'], nextport, name=shard)
                 nextport += 1
             elif self.args['replicaset']:
-                shard_names[p] = self._launchReplSet(self.args['dir'], nextport, shard, self.args['nodes'], self.args['arbiter'], self.args['verbose'])
+                shard_names[p] = self._launchReplSet(self.args['dir'], nextport, shard)
                 nextport += self.args['nodes']
                 if self.args['arbiter']:
                     nextport += 1
@@ -179,16 +176,20 @@ class MongoLauncher(object):
             config_names = ['config1', 'config2', 'config3']
             
         for name in config_names:
-            self._launchConfig(self.args['dir'], nextport, name, verbose=self.args['verbose'])
+            self._launchConfig(self.args['dir'], nextport, name)
             config_string.append('%s:%i'%(self.hostname, nextport))
             nextport += 1
         
-        # start up mongos
-        self._launchMongoS(os.path.join(self.args['dir'], 'data', 'mongos.log'), nextport, ','.join(config_string), auth=self.args['authentication'], loglevel=self.args['loglevel'], verbose=self.args['verbose'])
-        self.mongos_host = '%s:%i'%(self.hostname, nextport)
+        # start up mongos (at least 1)
+        for i in range(max(1, self.args['mongos'])):
+            self._launchMongoS(os.path.join(self.args['dir'], 'data', 'mongos.log'), nextport, ','.join(config_string))
+            if i == 0: 
+                # store host/port of first mongos
+                self.mongos_host = '%s:%i'%(self.hostname, nextport)
+            nextport += 1
 
         # add shards
-        print "adding shards (can take a few seconds) ..."
+        print "adding shards (can take a few seconds, grab a snickers) ..."
 
         con = Connection(self.mongos_host)
 
@@ -222,42 +223,38 @@ class MongoLauncher(object):
             time.sleep(1)
 
 
-
-
-    def _launchReplSet(self, basedir, portstart, name, numdata, arbiter, verbose=False):
+    def _launchReplSet(self, basedir, portstart, name):
         threads = []
         configDoc = {'_id':name, 'members':[]}
 
-        for i in range(numdata):
-            datapath = self._createPaths(basedir, '%s/rs%i'%(name, i+1), verbose)
-            self._launchMongoD(os.path.join(datapath, 'db'), os.path.join(datapath, 'mongod.log'), portstart+i, replset=name, auth=self.args['authentication'], loglevel=self.args['loglevel'], verbose=verbose)
+        for i in range(self.args['nodes']):
+            datapath = self._createPaths(basedir, '%s/rs%i'%(name, i+1))
+            self._launchMongoD(os.path.join(datapath, 'db'), os.path.join(datapath, 'mongod.log'), portstart+i, replset=name)
         
             host = '%s:%i'%(self.hostname, portstart+i)
             configDoc['members'].append({'_id':len(configDoc['members']), 'host':host})
             threads.append(threading.Thread(target=pingMongoDS, args=(host, 1.0, 30)))
-            if verbose:
+            if self.args['verbose']:
                 print "waiting for mongod at %s to start up..."%host
-
-            print "mongod at %s running." % host
 
         # launch arbiter if True
-        if arbiter:
-            datapath = self._createPaths(basedir, '%s/arb'%(name), verbose)
-            self._launchMongoD(os.path.join(datapath, 'db'), os.path.join(datapath, 'mongod.log'), portstart+numdata, replset=name, auth=self.args['authentication'], loglevel=self.args['loglevel'], verbose=verbose)
+        if self.args['arbiter']:
+            datapath = self._createPaths(basedir, '%s/arb'%(name))
+            self._launchMongoD(os.path.join(datapath, 'db'), os.path.join(datapath, 'mongod.log'), portstart+self.args['nodes'], replset=name)
             
-            host = '%s:%i'%(self.hostname, portstart+numdata)
+            host = '%s:%i'%(self.hostname, portstart+self.args['nodes'])
             configDoc['members'].append({'_id':len(configDoc['members']), 'host':host, 'arbiterOnly': True})
             threads.append(threading.Thread(target=pingMongoDS, args=(host, 1.0, 30)))
-            if verbose:
+            if self.args['verbose']:
                 print "waiting for mongod at %s to start up..."%host
-
-            print "arbiter at %s running." % host
 
         for thread in threads:
             thread.start()
 
         for thread in threads:
             thread.join()
+
+        print "all mongod processes for replica set '%s' running."%name
 
         # initiate replica set
         con = Connection('localhost:%i'%portstart)
@@ -266,33 +263,33 @@ class MongoLauncher(object):
             rs_status = con['admin'].command({'replSetGetStatus': 1})
         except OperationFailure, e:
             con['admin'].command({'replSetInitiate':configDoc})
-            if verbose:
+            if self.args['verbose']:
                 print "replica set configured."
 
         return name + '/' + ','.join([c['host'] for c in configDoc['members']])
 
 
-    def _launchConfig(self, basedir, port, name=None, verbose=False):
-        datapath = self._createPaths(basedir, name, verbose)
-        self._launchMongoD(os.path.join(datapath, 'db'), os.path.join(datapath, 'mongod.log'), port, replset=None, verbose=verbose, auth=self.args['authentication'], loglevel=self.args['loglevel'], extra='--configsvr')
+    def _launchConfig(self, basedir, port, name=None):
+        datapath = self._createPaths(basedir, name)
+        self._launchMongoD(os.path.join(datapath, 'db'), os.path.join(datapath, 'mongod.log'), port, replset=None, extra='--configsvr')
 
         host = '%s:%i'%(self.hostname, port)
         t = threading.Thread(target=pingMongoDS, args=(host, 1.0, 30))
         t.start()
-        if verbose:
+        if self.args['verbose']:
             print "waiting for mongod config server to start up..."
         t.join()
         print "mongod config server at %s running."%host
 
 
-    def _launchSingle(self, basedir, port, name=None, verbose=False):
-        datapath = self._createPaths(basedir, name, verbose)
-        self._launchMongoD(os.path.join(datapath, 'db'), os.path.join(datapath, 'mongod.log'), port, replset=None, auth=self.args['authentication'], loglevel=self.args['loglevel'], verbose=verbose)
+    def _launchSingle(self, basedir, port, name=None):
+        datapath = self._createPaths(basedir, name)
+        self._launchMongoD(os.path.join(datapath, 'db'), os.path.join(datapath, 'mongod.log'), port, replset=None)
 
         host = '%s:%i'%(self.hostname, port)
         t = threading.Thread(target=pingMongoDS, args=(host, 1.0, 30))
         t.start()
-        if verbose:
+        if self.args['verbose']:
             print "waiting for mongod to start up..."
         t.join()
         print "mongod at %s running."%host
@@ -300,64 +297,61 @@ class MongoLauncher(object):
         return host
 
 
-    def _launchMongoD(self, dbpath, logpath, port, replset=None, verbose=False, auth=False, loglevel=False, extra=''):
-        self.check_port_availability(port, "mongod")
-
+    def _launchMongoD(self, dbpath, logpath, port, replset=None, extra=''):
         rs_param = ''
         if replset:
             rs_param = '--replSet %s'%replset
 
         auth_param = ''
-        if auth:
+        if self.args['authentication']:
             key_path = os.path.abspath(os.path.join(self.args['dir'], 'data/keyfile'))
             auth_param = '--keyFile %s'%key_path
-
-        log_param = ''
-        if loglevel:
-            log_param = '-' + ''.join(['v']*loglevel)
 
         if self.args['rest']:
             extra = '--rest ' + extra
 
+        if self.unknown_args:
+            extra = ' '.join(self.unknown_args) + ' ' + extra
+
         local = ''
         if self.args['local']:
             local = "./"
 
-        ret = subprocess.call(['%smongod %s --dbpath %s --logpath %s --port %i --logappend %s %s %s --fork'%(local, rs_param, dbpath, logpath, port, auth_param, log_param, extra)], stderr=subprocess.STDOUT, stdout=subprocess.PIPE, shell=True)
-        if verbose:
-            print 'launching: %smongod %s --dbpath %s --logpath %s --port %i --logappend %s %s %s --fork'%(local, rs_param, dbpath, logpath, port, auth_param, log_param, extra)
+        ret = subprocess.call(['%smongod %s --dbpath %s --logpath %s --port %i --logappend %s %s --fork'%(local, rs_param, dbpath, logpath, port, auth_param, extra)], stderr=subprocess.STDOUT, stdout=subprocess.PIPE, shell=True)
+        if self.args['verbose']:
+            print 'launching: %smongod %s --dbpath %s --logpath %s --port %i --logappend %s %s --fork'%(local, rs_param, dbpath, logpath, port, auth_param, extra)
 
         return ret
 
 
-    def _launchMongoS(self, logpath, port, configdb, auth=False, loglevel=False, verbose=False):
-        self.check_port_availability(port, "mongos")
+    def _launchMongoS(self, logpath, port, configdb):
+        extra = ''
 
         out = subprocess.PIPE
-        if verbose:
+        if self.args['verbose']:
             out = None
 
         auth_param = ''
-        if auth:
+        if self.args['authentication']:
             key_path = os.path.abspath(os.path.join(self.args['dir'], 'data/keyfile'))
             auth_param = '--keyFile %s'%key_path
-
-        log_param = ''
-        if loglevel:
-            log_param = '-' + ''.join(['v']*loglevel)
 
         local = ''
         if self.args['local']:
             local = "./"
 
-        ret = subprocess.call(['%smongos --logpath %s --port %i --configdb %s --logappend %s %s --fork'%(local, logpath, port, configdb, auth_param, log_param)], stderr=subprocess.STDOUT, stdout=subprocess.PIPE, shell=True)
-        if verbose:
-            print 'launching: %smongos --logpath %s --port %i --configdb %s --logappend %s %s --fork'%(local, logpath, port, configdb, auth_param, log_param)
+        if self.unknown_args:
+            mongos_args = [arg for arg in self.unknown_args if arg in self.mongos_arguments]
+            extra = ' '.join(mongos_args) + extra
+
+        ret = subprocess.call(['%smongos --logpath %s --port %i --configdb %s --logappend %s %s --fork'%(local, logpath, port, configdb, auth_param, extra)], stderr=subprocess.STDOUT, stdout=subprocess.PIPE, shell=True)
+        if self.args['verbose']:
+            print 'launching: %smongos --logpath %s --port %i --configdb %s --logappend %s %s --fork'%(local, logpath, port, configdb, auth_param, extra)
         
         host = '%s:%i'%(self.hostname, port)
         t = threading.Thread(target=pingMongoDS, args=(host, 1.0, 30))
         t.start()
-        if verbose:
+        if self.args['verbose']:
             print "waiting for mongos to start up..."
         t.join()
         print "mongos at %s running."%host
@@ -365,5 +359,6 @@ class MongoLauncher(object):
 
 
 if __name__ == '__main__':
-    mongoLauncher = MongoLauncher()
+    tool = MLaunchTool()
+    tool.run()