+++ CHANGES.md
@@ -1,7 +1,19 @@
 Changes to mtools
 =================
 
-#### Version 1.0.0
+#### version 1.0.1
+
+  * fixed timezone bug in mlogmerge (#24)
+  * allow for multiple mongos in mlaunch with `--mongos` NUM parameter (#30)
+  * mlaunch can now take any additional arguments (like `-vvv`) and pass it on to the mongod/s instances (#31)
+  * all scripts now have --version flag (inherited from BaseCmdLineTool) (#34)
+  * added `--fast` option to mlogfilter (#37)
+  * fixed bug in logline parsing when detecting duration at the end of a line
+  * changed `--log` to `--logscale` argument for mplotqueries to avoid confusion with "log" files
+  * restructured tools to derive from base class `BaseCmdLineTool` or LogFileTool
+
+
+#### version 1.0.0
 
 This is the first version of mtools that has a version number. Some significant changes to its unnumbered predecessor are: